+++ Sources/ViewModels/ALKConversationViewModel.swift
@@ -571,10 +571,10 @@ open class ALKConversationViewModel: NSObject, Localizable {
         for message in messages {
             if channelKey != nil, channelKey == message.groupId {
                 filteredArray.append(message)
-                delegate?.updateTyingStatus(status: false, userId: message.to)
+//                delegate?.updateTyingStatus(status: false, userId: message.to)
             } else if message.channelKey == nil, channelKey == nil, contactId == message.to {
                 filteredArray.append(message)
-                delegate?.updateTyingStatus(status: false, userId: message.to)
+//                delegate?.updateTyingStatus(status: false, userId: message.to)
             }
         }