+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
 
+## [Unreleased]
+
+- Send Post Submitted Form Data As Message
+- Added Support for RTL Languages
+
 ## [0.1.1] - 2022-01-11
 - Fixed the Chat bar RTL support issue
 - Updated Core SDk to 0.0.2